requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
fonttools==4.54.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rich==13.9.4
rich-toolkit==0.13.2
scikit-image==0.25.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.29.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
gunicorn